# docs/swagger.yaml
definitions:
  types.ErrorObject:
    properties:
      error: {}
      message:
        type: string
      nonce:
        type: integer
      status:
        default: 400
        type: integer
    type: object
  types.FileDeleteJson:
    properties:
      mountPathWithName:
        type: string
    required:
    - mountPathWithName
    type: object
  types.ResponseObject:
    properties:
      message:
        type: string
      nonce:
        type: integer
      payload: {}
      status:
        default: 200
        type: integer
    type: object
info:
  contact: {}
paths:
  /files/compress:
    post:
      consumes:
      - multipart/form-data; boundary=normal
      description: execution will upload any file in compress format
      parameters:
      - default: false
        in: formData
        name: encodeWaiting
        required: true
        type: boolean
      - default: -filter:v fps=25 -vf scale=1280:720 -b:v 880k -b:a 128k -c:v h264
          -c:a aac -ac 2 -ar 44100
        in: formData
        name: ffmpegStr
        type: string
      - default: videoclips
        in: formData
        name: folder
        required: true
        type: string
      - default: /mnt/streamstudio
        in: formData
        name: mountPath
        required: true
        type: string
      - in: formData
        name: name
        type: string
      - default: mp4
        in: formData
        name: outputFormat
        type: string
      - description: File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ResponseObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorObject'
      summary: upload any file in compress format
      tags:
      - Files
  /files/raw:
    post:
      consumes:
      - multipart/form-data; boundary=normal
      description: execution will upload any file in raw format
      parameters:
      - default: videoclips
        in: formData
        name: folder
        required: true
        type: string
      - default: /mnt/streamstudio
        in: formData
        name: mountPath
        required: true
        type: string
      - in: formData
        name: name
        type: string
      - description: File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ResponseObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorObject'
      summary: upload any file in raw format
      tags:
      - Files
  /files/remove:
    patch:
      consumes:
      - application/json
      description: execution will delete a file from specific location
      parameters:
      - description: Mount Path With Name
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/types.FileDeleteJson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ResponseObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorObject'
      summary: delete a file from specific location
      tags:
      - Files
  /status:
    get:
      consumes:
      - application/json
      description: execution will return gin and cpu status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ResponseObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorObject'
      summary: returns gin and cpu status
      tags:
      - Status
swagger: "2.0"